:recycle: Refs. #235 -- refactor logic|price rule saving code
